Create pymol_COM.sh
This script converts all vina docked poses to pdb files and then reads all files using pymol externally. Find the center of mass for each pose and records the x,y,z coordinate as if a water molecule in a xyz file. It will be very useful to see how the COMs are distributed over the protein.